+++ lib/survey_kit.dart
@@ -82,7 +82,4 @@ export 'src/widget/survey_progress_configuration.dart';
 export 'src/widget/survey_progress.dart';
 export 'src/views/widget/survey_app_bar.dart';
 
-export 'src/steps/predefined_steps/video_step.dart';
-export 'src/views/video_step_view.dart';
-
 export 'src/configuration/app_bar_configuration.dart';